sumthemall

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -9,12 +9,12 @@ int sum_them_all(const unsigned int n, ...)
 	va_list ap;
 	int i, sum = 0;
 
-	if (n == 0)
+	if ((int)n == 0)
 	{
 		return (0);
 	}
 	va_start(ap, n);
-	for (i = 0; i < n; i++)
+	for (i = 0; i < (int)n; i++)
 	{
 		sum += va_arg(ap, int);
 	}
